% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_node_deps}
\alias{check_node_deps}
\title{Check node module dependencies and install if missing}
\usage{
check_node_deps(node_package, r_package, node_dir = "node")
}
\arguments{
\item{node_package}{the directory name of the node package in the R package}

\item{r_package}{the package name which wraps the function. Must be specified
as the function can't automatically detect the package
inside \link{.onAttach}}

\item{node_dir}{the directory where node packages are kept.  Defaults to
'node', which should be a directory under 'inst' when
creaing your own package}
}
\description{
Put this function in \link{.onAttach} to check and install node module
dependencies on first load.  This is useful because some dependencies are
binary and can not be hosted on CRAN.  Note that it is also run each time a
function created by \link{node_fn} is called.
}

